# [DO NOT MERGE] API Removed SS_ConfigStaticManifest in favour of Reflection