--- a/composer.json
+++ b/composer.json
@@ -28,7 +28,7 @@
 		}
 	},
 	"autoload": {
-		"classmap": ["tests/behat/features/bootstrap"]	
+		"classmap": ["tests/behat/features/bootstrap"]
 	},
 	"minimum-stability": "dev"
 }