Apologies for the "do not merge", but this is the best way to show a diff and include as many people as possible.

Anyway, i'm hoping that this PR is the start of a discussion to remove SS_ConfigStaticManifest and its flush process.

Instead of collecting private statics through parsing PHP files during the flush process, this PR uses Reflection to get private static config values at run-time when they're requested. This doesn't remove or affect YAML configuration in any way.

Initial performance tests show a small speed increase and ~0.6mb memory saving:

Before:

	Parameter			Min.		Max.		Avg.
	CPU Time			85.104 ms	180.61 ms	94.278 ms
	Memory Usage		9.97 M		10 M		9.98 M
	Peak Memory Usage	10.07 M		10.1 M		10.07 M
	Dataset Size		50

After:

	Parameter			Min.		Max.		Avg.
	CPU Time			85.052 ms	181.952 ms	92.791 ms
	Memory Usage		9.35 M		9.38 M		9.36 M
	Peak Memory Usage	9.34 M		9.49 M		9.46 M
	Dataset Size		50

These were ran using php 5.5.

This is only a proof of concept, but if nobody raises any issues with this method i'll turn this into a proper PR. Feedback welcome.