@dhensby 2015-03-23
We're getting a rather strange test failure and I'd expect we aren't seeing more as the test coverage is probably not as good as it should be for this part of the system.