@dhensby 2015-03-23
Please can you elaborate on how you tested this and came to conclude it was faster and used less memory? Did you test on a production site (ie: with lots of functionality) or on a clean install?