@micmania1 2015-03-23
I tested using a base install of master on PHP 5.5 using xhprof.

The goal of this wasn't actually to improve performance initially - the test was to ensure that this doesn't introduce any new performance issues. The memory saving was expected but also I would expect the flush process to take less time. 

The long goal is to improve performance but the initial goal is to simplify things - parsing PHP files to get static values seems a bit OTT.

The test failures make sense. This solution is case-sensitive so that's something I'd need to consider in the final solution. I haven't looked into how the case-insensitivity is currently implemented so that'll be next on my list :)