@kinglozzer 2015-04-01
Alternative suggestion (5.4+ only, though): [`Closure::bind()`](http://php.net/manual/en/closure.bind.php).

```php
// Note: actual code would store each $thief (as one is needed for each class)
$thief = function($name) {
	return self::$$name;
};

$thief = Closure::bind($thief, null, $staticSource);
$value = $thief($name);
```

Just had a quick play with it, seems to be slightly (~3%) more performant than `ReflectionProperty`. Uses ~0.3% more memory than this solution (~66kb more), but still uses less memory than the existing static manifests. Only benchmarked on one site - I’ll play around with some that vary in complexity :)