@chillu 2015-04-27
Given the complexity and inherent brittleness of `SS_ConfigStaticManifest_Parser`, I'm generally in favour of this change. We'd no longer be at the mercy of a correctly implemented regex which can parse all possible syntax variations. It'll also open up more possibilities around annotations.

I have no strong opinions on `Reflection` vs. `Closure::bind()`.

Have you tested on an SSD? I'm not sure how PHP 5.5's built-in opcache works, i.e. if `scandir()` and `file_get_contents()` in `ManifestFileFinder` incur a higher performance penalty on slower disks.

I guess there's no need to test this against PHP versions older than 5.5 (with potentially slower Reflection implementations), since a 4.x release is unlikely to support them.